typescriptify for better maintainability
